Variance Explained Plots for MDIVI
There is an error in numbering. Figure 5 should actually be titled Figure 4. Figure 6 should be Figure 5 and so on. Figures 1 to 3 are not affected by this numbering error. 

NOTE: Figure 10 and Figure 17 are both BLANK. Revisiting the original MDIVI eigenvalue arrays that these figures corresponded to revealed that these two arrays are just composed of 0s. These problematic arrays are DsRed2_HeLa_11_21_4 (figure 10) and DsRed2_HeLa_11_21_12 (figure 17).